feat: separate lora transform from quantization
